game controller refactored
