+++ New-Model/proj/src/controller/gameControllers/KingdomController.java
@@ -1,0 +1,28 @@
+package controller.gameControllers;
+
+import java.util.regex.Matcher;
+
+public class KingdomController {
+    GameMenuController gameMenuController;
+
+    public KingdomController(GameMenuController gameMenuController) {
+        this.gameMenuController = gameMenuController;
+    }
+    public String changeFoodRate(Matcher matcher) {
+        //return a proper response if there is a problem
+        //change it if there is no problem
+        return "";
+    }
+
+    public String changeTaxRate(Matcher matcher) {
+        //return a proper response if there is a problem
+        //change it if there is no problem
+        return "";
+    }
+
+    public String changeFearRate(Matcher matcher) {
+        //return a proper response if there is a problem
+        //change it if there is no problem
+        return "";
+    }
+}
