+++ New-Model/proj/src/controller/gameControllers/MapController.java
@@ -1,0 +1,19 @@
+package controller.gameControllers;
+
+import java.util.regex.Matcher;
+
+public class MapController {
+    GameMenuController gameMenuController;
+    public MapController(GameMenuController gameMenuController) {
+        this.gameMenuController = gameMenuController;
+    }
+    public void showMap(Matcher matcher) {
+
+    }
+    public void moveMap(Matcher matcher) {
+
+    }
+    public void showDetails(Matcher matcher) {
+
+    }
+}
