+++ New-Model/proj/src/controller/gameControllers/TradeController.java
@@ -1,0 +1,25 @@
+package controller.gameControllers;
+
+import java.util.regex.Matcher;
+public class TradeController {
+    GameMenuController gameMenuController;
+
+    public TradeController(GameMenuController gameMenuController) {
+        this.gameMenuController = gameMenuController;
+    }
+    public String requestTrade(Matcher matcher) {
+        return "";
+    }
+
+    public String showTradeList() {
+        return "";
+    }
+
+    public String acceptTrade() {
+        return "";
+    }
+
+    public String showTradeHistory() {
+        return "";
+    }
+}
