+++ New-Model/proj/src/view/GameMenu.java
@@ -1,7 +1,5 @@
 package view;
 
-import model.Field;
-
 import java.util.Scanner;
 import java.util.regex.Matcher;
